backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend API for organizing collective recitation of Quran and Islamic texts"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pyjwt>=2.9.0",
    "pydantic-settings>=2.6.0",
    "pydantic[email]>=2.12.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]
